requirements.txt
fastapi
httpx[http2]
uvicorn[standard]
azure-identity
python-dotenv
